+++ Frontend/healthcare/src/Pages/Staff/ViewStaff.jsx
@@ -5,6 +5,7 @@ import Sidebar from '../../Components/sidebar';
 import { makeStyles } from '@material-ui/core/styles';
 import { Edit, Delete } from '@material-ui/icons';
 import { useNavigate } from 'react-router-dom';
+import swal from 'sweetalert2';
 
 // Custom Pagination Component
 const CustomPagination = ({ count, page, rowsPerPage, onPageChange }) => {
@@ -89,13 +90,28 @@ const ViewStaff = () => {
   };
 
   const handleDelete = async (id) => {
-    try {
-      await axios.delete(`http://localhost:3002/staff/delete-staff/${id}`);
-      setStaffData(staffData.filter(staff => staff._id !== id));
-    } catch (error) {
-      console.error("There was an error deleting the staff member!", error);
+    const result = await swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this staff record!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    });
+  
+    if (result.isConfirmed) {
+      try {
+        await axios.delete(`http://localhost:3002/staff/delete-staff/${id}`);
+        setStaffData(staffData.filter(staff => staff._id !== id));
+        swal.fire('Deleted!', 'The record has been deleted.', 'success');
+      } catch (error) {
+        console.error("There was an error deleting the staff member!", error);
+        swal.fire('Error!', 'There was an error deleting the record.', 'error');
+      }
     }
   };
+  
 
   const handleSearchQueryChange = (event) => {
     setSearchQuery(event.target.value);
@@ -177,7 +193,7 @@ const ViewStaff = () => {
                     <TableCell style={{ color: 'white' }}>Email</TableCell>
                     <TableCell style={{ color: 'white' }}>Position</TableCell>
                     <TableCell style={{ color: 'white' }}>Department</TableCell>
-                    <TableCell style={{ color: 'white' }}>Date of Joining</TableCell>
+                    <TableCell style={{ color: 'white' }}>Joined</TableCell>
                     <TableCell style={{ color: 'white' }}>Status</TableCell>
                     <TableCell style={{ color: 'white' }}>Update</TableCell>
                     <TableCell style={{ color: 'white' }}>Delete</TableCell>
